Bonus problem solution
Hello,

I've noticed that the current solution to the bonus problem is making comparisons between distinct countries. For instance, row 13 (Albania 1952) is being compared to row 12 (Afghanistan 2007) and then being incorrectly included in the answer. Also, it is assuming that the data are ordered by countries and years (indeed it is, but it is not safe to assume so).

In this patch I propose a solution to both issues.

Kind regards,
Romero Morais